Do not drop roles tables in test cases
The image used for testing already has the roles table with data. If the
data is dropped and new roles are created, then their id won't have the
expected value.

+++ tests/integration_tests/framework/postgresql.py
@@ -48,11 +48,13 @@ def run_query(query, db_name=None):
             return {'status': status_message, 'all': fetchall}
 
 
-def safe_drop_all():
+def safe_drop_all(keep_tables):
     """Creates a single transaction that *always* drops all tables, regardless
     of relationships and foreign key constraints (as opposed to `db.drop_all`)
     """
     meta = db.metadata
     for table in reversed(meta.sorted_tables):
+        if table.name in keep_tables:
+            continue
         db.session.execute(table.delete())
     db.session.commit()
